add registger and login functionality
