+++ src/components/unauthenticated/auth_forms/AuthFormsStyles.js
@@ -1,1 +1,21 @@
 import styled from 'styled-components'
+// export const asdf = styled.``
+export const AuthHeader = styled.h2`
+  color: red;
+`
+
+export const AuthView = styled.div`
+  /* border: 3px solid purple; */
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  p a {
+    text-decoration: none;
+    color: #0467fb;
+    font-weight: bold;
+    :hover {
+      color: #4b59f7;
+    }
+  }
+`
